Finish v7.1.3 Incident Command Wire controls and alignment
