--- .github/workflows/apply-issue519-v713-final-ui.yml
@@ -1,133 +1,0 @@
-name: Build Issue 519 v7.1.3 Final Wire UI
-
-on:
-  pull_request:
-    branches: [main]
-    paths:
-      - ".github/issue519/apply_v713_final_ui.py"
-      - ".github/workflows/apply-issue519-v713-final-ui.yml"
-
-permissions:
-  contents: read
-
-concurrency:
-  group: toolkit-issue-519-v713-final-ui
-  cancel-in-progress: false
-
-jobs:
-  build:
-    if: github.actor == 'Conroy1988' && github.head_ref == 'fix/issue-519-v713-reduced-motion-autoplay'
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    steps:
-      - name: Check out exact v7.1.3 branch
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          ref: fix/issue-519-v713-reduced-motion-autoplay
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Remove nonexistent standalone priority-badge edit
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 - <<'PY'
-          import re
-          from pathlib import Path
-          path = Path('.github/issue519/apply_v713_final_ui.py')
-          text = path.read_text(encoding='utf-8')
-          pattern = re.compile(
-              r'\n    source = append_css_properties\(\n'
-              r'        source,\n'
-              r"        '#\$\{SCRIPT\.majorIncidentFeedId\} \.mcms-incident-level',\n"
-              r"        'align-self:center!important;line-height:1!important;margin-block:0!important;transform:none!important;',\n"
-              r'        "priority badge alignment",\n'
-              r'    \)\n'
-          )
-          text, count = pattern.subn('\n', text, count=1)
-          if count != 1:
-              raise SystemExit(f'Expected one standalone priority-badge edit, found {count}')
-          path.write_text(text, encoding='utf-8')
-          PY
-
-      - name: Apply dropdown-only and vertical-centre cleanup
-        id: apply
-        shell: bash
-        run: |
-          set +e
-          set -o pipefail
-          python3 .github/issue519/apply_v713_final_ui.py 2>&1 | tee issue519-v713-final-ui.log
-          STATUS=${PIPESTATUS[0]}
-          echo "exit_code=$STATUS" >> "$GITHUB_OUTPUT"
-          exit 0
-
-      - name: Upload final UI transformation diagnostic
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: issue519-v713-final-ui-${{ github.sha }}
-          path: issue519-v713-final-ui.log
-          if-no-files-found: error
-          retention-days: 14
-
-      - name: Stop on transformation failure
-        if: steps.apply.outputs.exit_code != '0'
-        run: exit 1
-
-      - name: Scope runtime control assertions to rendered template
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 - <<'PY'
-          from pathlib import Path
-          path = Path('.github/scripts/test_issue517_incident_command_wire_runtime.js')
-          text = path.read_text(encoding='utf-8')
-          old = '''assert.equal(source.includes('data-mcms-incident-action="previous"'),false);assert.equal(source.includes('data-mcms-incident-action="pause"'),false);assert.equal(source.includes('data-mcms-incident-action="next"'),false);assert.equal(source.includes('data-mcms-incident-action="expand"'),true);'''
-          new = '''const ensureStart=source.indexOf('    function ensureMajorIncidentFeed(');const renderStart=source.indexOf('    function renderMajorIncidentFeed(',ensureStart);const ensure=source.slice(ensureStart,renderStart);const templateStart=ensure.indexOf('feed.innerHTML = `');const templateEnd=ensure.indexOf('`;',templateStart);const template=ensure.slice(templateStart,templateEnd);assert.equal(template.includes('data-mcms-incident-action="previous"'),false);assert.equal(template.includes('data-mcms-incident-action="pause"'),false);assert.equal(template.includes('data-mcms-incident-action="next"'),false);assert.equal(template.includes('data-mcms-incident-action="expand"'),true);'''
-          if text.count(old) != 1:
-              raise SystemExit(f'Expected one unscoped runtime assertion chain, found {text.count(old)}')
-          path.write_text(text.replace(old, new, 1), encoding='utf-8')
-          PY
-
-      - name: Validate complete candidate
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 .github/scripts/validate_userscript.py
-          bash .github/scripts/run_userscript_preflight.sh --contracts
-          python3 .github/scripts/test_documentation_consistency.py
-          node --check src/MissionChief_Map_Command_Toolkit.user.js
-          cmp --silent dist/MissionChief_Map_Command_Toolkit.user.js dist/MissionChief_Map_Command_Toolkit.txt
-
-      - name: Enforce unchanged performance envelope
-        shell: bash
-        run: |
-          set -euo pipefail
-          git fetch origin main --depth=1
-          git show origin/main:src/MissionChief_Map_Command_Toolkit.user.js > /tmp/performance-base.user.js
-          python3 .github/scripts/test_performance_budget.py
-          python3 .github/scripts/check_performance_budget.py \
-            --candidate src/MissionChief_Map_Command_Toolkit.user.js \
-            --base /tmp/performance-base.user.js \
-            --policy .github/performance-budget.json \
-            --json-output /tmp/issue519-v713-final-ui.json \
-            --markdown-output /tmp/issue519-v713-final-ui.md
-          cat /tmp/issue519-v713-final-ui.md
-
-      - name: Commit validated final UI candidate
-        env:
-          DEVELOPMENT_PR_TOKEN: ${{ secrets.DEVELOPMENT_PR_TOKEN }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          test -n "$DEVELOPMENT_PR_TOKEN"
-          test "$GITHUB_HEAD_REF" = "fix/issue-519-v713-reduced-motion-autoplay"
-          rm -f issue519-v713-final-ui.log
-          git config user.name "Conroy1988"
-          git config user.email "27301455+Conroy1988@users.noreply.github.com"
-          git add -A
-          git diff --cached --quiet && { echo "No final Incident Wire UI changes produced."; exit 1; }
-          git commit -m "Finish v7.1.3 Incident Command Wire controls and alignment"
-          BASIC_AUTH="$(printf 'x-access-token:%s' "$DEVELOPMENT_PR_TOKEN" | base64 -w0)"
-          git -c http.https://github.com/.extraheader="AUTHORIZATION: basic ${BASIC_AUTH}" \
-            push origin HEAD:refs/heads/fix/issue-519-v713-reduced-motion-autoplay
