+++ src/MissionChief_Map_Command_Toolkit.user.js
@@ -8994,19 +8994,19 @@ html[data-mc-map-skin="default"] .leaflet-tile-pane img.leaflet-tile { filter: n
         
         /* v7.1.0 Incident Command Wire: bounded, theme-aware card navigation. */
         #${SCRIPT.majorIncidentFeedId}{--mcms-wire-accent:#68cfff;--mcms-wire-border:#467ca4;--mcms-wire-bg:#0a161f;--mcms-wire-bg2:#142734;--mcms-wire-label:linear-gradient(180deg,#24628e,#173e5b);--mcms-wire-text:#f4fbff;--mcms-wire-muted:#abc0cd;position:relative;height:44px!important;overflow:visible!important;border-color:var(--mcms-wire-border)!important;background:linear-gradient(180deg,var(--mcms-wire-bg2),var(--mcms-wire-bg))!important;color:var(--mcms-wire-text)!important}
-        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-label{min-width:148px!important;padding:0 9px!important;gap:7px!important;border-right:1px solid var(--mcms-wire-border)!important;background:var(--mcms-wire-label)!important}
-        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-label-title,#${SCRIPT.majorIncidentFeedId} .mcms-incident-name,#${SCRIPT.majorIncidentFeedId} .mcms-incident-state{overflow:hidden;text-overflow:ellipsis;white-space:nowrap}
-        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-count{flex:0 0 auto;min-width:34px;padding:3px 5px;border:1px solid currentColor;border-radius:999px;font-size:7px;line-height:1;text-align:center}
+        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-label{min-width:148px!important;padding:0 9px!important;gap:7px!important;border-right:1px solid var(--mcms-wire-border)!important;background:var(--mcms-wire-label)!importantalign-items:center!important;line-height:1!important;}
+        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-label-title,#${SCRIPT.majorIncidentFeedId} .mcms-incident-name,#${SCRIPT.majorIncidentFeedId} .mcms-incident-state{overflow:hidden;text-overflow:ellipsis;white-space:nowrapalign-self:center!important;line-height:1!important;margin-block:0!important;transform:none!important;}
+        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-count{flex:0 0 auto;min-width:34px;padding:3px 5px;border:1px solid currentColor;border-radius:999px;font-size:7px;line-height:1;text-align:centeralign-self:center!important;margin-block:0!important;transform:none!important;}
         #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-viewport{height:100%!important;min-width:0;overflow:hidden!important}
         #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-track{display:flex!important;align-items:stretch!important;width:max-content!important;min-width:200%!important;height:100%!important;will-change:transform;animation:mcmsIncidentWireReel var(--mcms-incident-feed-duration,90s) linear infinite!important;transition:none!important}
         @keyframes mcmsIncidentWireReel{from{transform:translate3d(0,0,0)}to{transform:translate3d(-50%,0,0)}}
-        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-group{display:flex!important;align-items:stretch!important;flex:0 0 auto!important;width:max-content!important;min-width:max-content!important;height:100%!important}#${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-item{flex:0 0 auto!important;width:clamp(520px,52vw,820px)!important;min-width:clamp(520px,52vw,820px)!important;max-width:820px!important;height:100%!important;padding:0 12px!important;gap:9px!important;overflow:hidden!important;border:0!important;border-left:3px solid var(--mcms-wire-accent)!important;background:linear-gradient(90deg,rgba(104,207,255,.08),transparent 30%)!important}#${SCRIPT.majorIncidentFeedId}.mcms-feed-paused .mcms-incident-feed-track,#${SCRIPT.majorIncidentFeedId}.mcms-feed-interacting .mcms-incident-feed-track,#${SCRIPT.majorIncidentFeedId}.mcms-feed-expanded .mcms-incident-feed-track,#${SCRIPT.majorIncidentFeedId}.mcms-feed-static .mcms-incident-feed-track{animation-play-state:paused!important}
-        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-copy,#${SCRIPT.majorIncidentFeedId} .mcms-incident-meta{min-width:0;display:flex;align-items:center;gap:7px;overflow:hidden}
+        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-group{display:flex!important;align-items:stretch!important;flex:0 0 auto!important;width:max-content!important;min-width:max-content!important;height:100%!important}#${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-item{flex:0 0 auto!important;width:clamp(520px,52vw,820px)!important;min-width:clamp(520px,52vw,820px)!important;max-width:820px!important;height:100%!important;padding:0 12px!important;gap:9px!important;overflow:hidden!important;border:0!important;border-left:3px solid var(--mcms-wire-accent)!important;background:linear-gradient(90deg,rgba(104,207,255,.08),transparent 30%)!importantdisplay:flex!important;align-items:center!important;justify-content:flex-start!important;line-height:1!important;padding-top:0!important;padding-bottom:0!important;vertical-align:middle!important;}#${SCRIPT.majorIncidentFeedId}.mcms-feed-paused .mcms-incident-feed-track,#${SCRIPT.majorIncidentFeedId}.mcms-feed-interacting .mcms-incident-feed-track,#${SCRIPT.majorIncidentFeedId}.mcms-feed-expanded .mcms-incident-feed-track,#${SCRIPT.majorIncidentFeedId}.mcms-feed-static .mcms-incident-feed-track{animation-play-state:paused!important}
+        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-copy,#${SCRIPT.majorIncidentFeedId} .mcms-incident-meta{min-width:0;display:flex;align-items:center;gap:7px;overflow:hiddenalign-self:center!important;height:100%!important;line-height:1!important;margin-block:0!important;transform:none!important;}
         #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-copy{flex:1 1 auto}.mcms-incident-feed-meta{color:var(--mcms-wire-muted)}
         #${SCRIPT.majorIncidentFeedId} .mcms-incident-name{min-width:80px;max-width:52%;color:var(--mcms-wire-text)!important}
         #${SCRIPT.majorIncidentFeedId} .mcms-incident-meta{color:var(--mcms-wire-muted)!important}#${SCRIPT.majorIncidentFeedId} .mcms-incident-postcode{color:var(--mcms-wire-accent)!important}
         #${SCRIPT.majorIncidentFeedId} .mcms-incident-state{flex:0 0 auto;max-width:180px}
-        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-controls{position:relative!important;inset:auto!important;z-index:7;flex:0 0 auto;align-self:center!important;display:flex!important;align-items:center!important;justify-content:center!important;height:calc(100% - 4px)!important;min-height:0!important;max-height:36px!important;box-sizing:border-box!important;margin:0!important;gap:3px;padding:0 5px!important;border-left:1px solid var(--mcms-wire-border);background:rgba(0,0,0,.12);overflow:hidden!important;transform:none!important;contain:layout paint}
+        #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-controls{position:relative!important;inset:auto!important;z-index:7;flex:0 0 auto;align-self:center!important;display:flex!important;align-items:center!important;justify-content:center!important;height:calc(100% - 4px)!important;min-height:0!important;max-height:36px!important;box-sizing:border-box!important;margin:0!important;gap:3px;padding:0 5px!important;border-left:1px solid var(--mcms-wire-border);background:rgba(0,0,0,.12);overflow:hidden!important;transform:none!important;contain:layout paintwidth:38px!important;min-width:38px!important;max-width:38px!important;flex-basis:38px!important;padding:0 4px!important;}
         #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-controls button{appearance:none!important;-webkit-appearance:none!important;position:relative!important;inset:auto!important;float:none!important;align-self:center!important;box-sizing:border-box!important;display:grid!important;place-items:center!important;width:26px!important;height:26px!important;min-width:26px!important;min-height:26px!important;max-width:26px!important;max-height:26px!important;margin:0!important;padding:0!important;border:1px solid var(--mcms-wire-border)!important;border-radius:5px!important;background:rgba(255,255,255,.055);color:var(--mcms-wire-text);font:900 13px/1 Arial,sans-serif!important;cursor:pointer;overflow:hidden!important;transform:none!important;vertical-align:middle!important}
         #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-controls button:hover,#${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-controls button:focus-visible,#${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-controls [aria-pressed="true"]{background:var(--mcms-wire-accent);color:#09131a;outline:1px solid var(--mcms-wire-accent)}
         #${SCRIPT.majorIncidentFeedId} .mcms-incident-feed-panel{position:absolute;z-index:12;top:calc(100% + 5px);right:0;width:min(860px,calc(100vw - 18px));max-height:min(58vh,520px);overflow:hidden;border:1px solid var(--mcms-wire-border);border-radius:9px;background:linear-gradient(180deg,var(--mcms-wire-bg2),var(--mcms-wire-bg));box-shadow:0 15px 38px rgba(0,0,0,.52);color:var(--mcms-wire-text)}
@@ -14515,9 +14515,9 @@ The sweep opens verified alliance-owned FMS 5 patient vehicles and uses MissionC
             feed.innerHTML = `<div class="mcms-incident-feed-label"><span class="mcms-incident-feed-label-title"></span><span class="mcms-incident-feed-count">0</span></div>
                 <div class="mcms-incident-feed-viewport" aria-live="polite"><div class="mcms-incident-feed-track"></div></div>
                 <div class="mcms-incident-feed-controls" aria-label="Incident wire controls">
-                    <button type="button" data-mcms-incident-action="previous" aria-label="Previous incident" title="Previous incident">‹</button>
-                    <button type="button" data-mcms-incident-action="pause" aria-label="Pause incident rotation" title="Pause incident rotation">Ⅱ</button>
-                    <button type="button" data-mcms-incident-action="next" aria-label="Next incident" title="Next incident">›</button>
+                    
+                    
+                    
                     <button type="button" data-mcms-incident-action="expand" aria-label="Expand incident queue" aria-expanded="false" title="Expand incident queue">⌄</button>
                 </div>
                 <div class="mcms-incident-feed-panel" hidden><div class="mcms-incident-feed-panel-head"><strong>Active major incidents</strong><span>Priority ordered</span></div><div class="mcms-incident-feed-list"></div></div>`;
